requirements/travis-ci.txt
django-discover-runner
mock
